refactor(delta): resolve source identities from the kernel schema
Move the Delta producer out of core/schema-identity and into the Delta
connector, and walk Delta Kernel's StructType instead of re-reading the
schema JSON with Jackson.

Delta schema shape is now interpreted in exactly one place. Structural
faults are reported by the kernel's own parse, so the resolver validates
only the column-mapping metadata layered on that shape; empty field
names still need an explicit check, since the kernel accepts a field no
path can address.

core/schema-identity keeps only the format-neutral model and no longer
declares any runtime dependency, so a per-format producer supplies
whatever native identity its source governs without the shared model
depending on that format.
