+++ connectors/catalogs/delta/pom.xml
@@ -61,6 +61,11 @@
       <artifactId>floecat-storage-spi</artifactId>
       <version>${project.version}</version>
     </dependency>
+    <dependency>
+      <groupId>ai.floedb.floecat</groupId>
+      <artifactId>floecat-schema-identity</artifactId>
+      <version>${project.version}</version>
+    </dependency>
 
     <dependency>
       <groupId>io.delta</groupId>
