+++ connectors/catalogs/delta/src/test/java/ai/floedb/floecat/connector/delta/identity/DeltaResolvedSchemaTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package ai.floedb.floecat.schema.identity.delta;
+package ai.floedb.floecat.connector.delta.identity;
 
 import static org.assertj.core.api.Assertions.assertThat;
